pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "channels_postgres"
version = "1.0.7"
description = "PostgreSQL-backed ASGI channel layer implementation"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "BSD" }
authors = [
  { name = "Daniel Osaetin", email = "f805nqs6j@relay.firefox.com" }
]
dependencies = [
  "msgpack>=1.0.7,<2",
  "asgiref>=3.7.2,<4",
  "channels>=4.0.0,<5",
  "psycopg>=3,<4"
]

[project.urls]
Homepage = "http://github.com/danidee10/channels_postgres/"

[project.optional-dependencies]
cryptography = [
  "cryptography>=44.0.2,<45"
]
tests = [
  "async-timeout",
  "cryptography>=44.0.2,<45",
  "mypy>=1.15.0,<2",
  "pylint>=3.3.6,<4",
  "pylint-django>=2.6.1,<3",
  "pytest>=8.3.5,<9",
  "pytest-asyncio>=0.26.0,<1",
  "pytest-cov>=6.1.1,<7",
  "ruff>=0.11.4,<1"
]


[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
exclude = ["tests"]

[tool.wheel]
universal = true

[tool.pylint.MAIN]
ignore-paths = ["channels_postgres/migrations/*"]

[tool.pytest.ini_options]
addopts = "-p no:django tests/ --cov=channels_postgres  --cov-branch --cov-report xml --cov-report html"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "module"

[tool.ruff]
line-length = 100

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
docstring-code-format = true

[tool.mypy]
exclude = ["channels_postgres/migrations/*"]
strict = true

[[tool.mypy.overrides]]
module = ["django.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["channels.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["psycopg.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["pytest.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["asgiref.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["msgpack.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["cryptography.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["async_timeout.*"]
ignore_missing_imports = true
